refactor(data-in-pipeline): move transformer => app
