+++ data-in-pipeline/pyproject.toml
@@ -17,6 +17,3 @@ api = { workspace = true }
 
 [dependency-groups]
 dev = ["pytest>=8.4.2"]
-
-[tool.pytest.ini_options]
-pythonpath = ["."]
